[vm] move bounds checking to top-level module
All other checks have been moved to the bytecode verifier.
