--- language/vm/src/checks.rs
@@ -1,6 +1,0 @@
-// Copyright (c) The Libra Core Contributors
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod bounds;
-
-pub use bounds::{BoundsCheck, BoundsChecker};
